UAR-787-ind-entities-involved-in-trust
