+++ src/controllers/update/update.trusts.individuals.or.entities.involved.controller.ts
@@ -1,26 +1,10 @@
-import { NextFunction, Request, Response } from "express";
-import { logger } from "../../utils/logger";
-import * as config from "../../config";
+import { NextFunction, Request, Response } from 'express';
+import { getTrustInvolvedPage, postTrustInvolvedPage } from "../../utils/trust/trust.involved";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
-  try {
-    logger.debugRequest(req, `${req.method} ${req.route.path}`);
-
-    return res.render(config.UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_PAGE, {
-      backLinkUrl: config.UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL,
-      templateName: config.UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_PAGE
-    });
-  } catch (error) {
-    next(error);
-  }
+  getTrustInvolvedPage(req, res, next, true);
 };
 
 export const post = (req: Request, res: Response, next: NextFunction) => {
-  try {
-    logger.debugRequest(req, `${req.method} ${req.route.path}`);
-
-    return res.redirect(config.UPDATE_TRUSTS_ASSOCIATED_WITH_THE_OVERSEAS_ENTITY_URL);
-  } catch (error) {
-    next(error);
-  }
+  postTrustInvolvedPage(req, res, next, true);
 };
