+++ src/controllers/update/update.trusts.tell.us.about.it.controller.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 
-import { getTrustDetails, postTrustDetails } from "../../../src/utils/trust.details";
+import { getTrustDetails, postTrustDetails } from "../../utils/trust.details";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
   getTrustDetails(req, res, next, true);
