+++ test/controllers/update/update.trusts.individuals.or.entities.involved.controller.spec.ts
@@ -4,6 +4,7 @@ jest.mock('../../../src/utils/application.data');
 jest.mock('../../../src/middleware/authentication.middleware');
 jest.mock('../../../src/middleware/company.authentication.middleware');
 jest.mock('../../../src/middleware/service.availability.middleware');
+jest.mock('../../../src/utils/trust/common.trust.data.mapper');
 
 import { beforeEach, jest, test, describe } from '@jest/globals';
 import request from 'supertest';
@@ -13,7 +14,7 @@ import app from '../../../src/app';
 import {
   UPDATE_TRUSTS_ASSOCIATED_WITH_THE_OVERSEAS_ENTITY_URL,
   UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL,
-  UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL,
+  TRUST_INVOLVED_URL
 } from '../../../src/config';
 import { authentication } from '../../../src/middleware/authentication.middleware';
 import { companyAuthentication } from '../../../src/middleware/company.authentication.middleware';
@@ -21,9 +22,9 @@ import { serviceAvailabilityMiddleware } from '../../../src/middleware/service.a
 import { getApplicationData } from '../../../src/utils/application.data';
 import { isActiveFeature } from '../../../src/utils/feature.flag';
 
-import { APPLICATION_DATA_MOCK } from '../../__mocks__/session.mock';
-import { PAGE_TITLE_ERROR, PAGE_NOT_FOUND_TEXT } from '../../__mocks__/text.mock';
-import { saveAndContinueButtonText } from '../../__mocks__/save.and.continue.mock';
+import { APPLICATION_DATA_MOCK, TRUST_WITH_ID } from '../../__mocks__/session.mock';
+import { PAGE_TITLE_ERROR, PAGE_NOT_FOUND_TEXT, TRUST_INVOLVED_TITLE } from '../../__mocks__/text.mock';
+import { mapCommonTrustDataToPage } from '../../../src/utils/trust/common.trust.data.mapper';
 
 const mockGetApplicationData = getApplicationData as jest.Mock;
 mockGetApplicationData.mockReturnValue( APPLICATION_DATA_MOCK );
@@ -40,6 +41,11 @@ mockServiceAvailabilityMiddleware.mockImplementation((req: Request, res: Respons
 const mockIsActiveFeature = isActiveFeature as jest.Mock;
 mockIsActiveFeature.mockReturnValue(true);
 
+const mockMapCommonTrustDataToPage = mapCommonTrustDataToPage as jest.Mock;
+
+const trustId = TRUST_WITH_ID.trust_id;
+const pageUrl = UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL + "/" + trustId + TRUST_INVOLVED_URL;
+
 describe('Update - Trusts - Individuals or entities involved', () => {
   beforeEach(() => {
     jest.clearAllMocks();
@@ -47,21 +53,24 @@ describe('Update - Trusts - Individuals or entities involved', () => {
 
   describe('GET tests', () => {
     test('when feature flag is on, page is returned', async () => {
+      const mockTrustData = {
+        trustName: 'dummy',
+      };
+      mockMapCommonTrustDataToPage.mockReturnValue(mockTrustData);
       mockIsActiveFeature.mockReturnValue(true);
 
-      const resp = await request(app).get(UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL);
+      const resp = await request(app).get(pageUrl);
 
       expect(resp.status).toEqual(200);
-      expect(resp.text).toContain('Individuals or entities involved in the trust');
-      expect(resp.text).toContain(UPDATE_TRUSTS_TELL_US_ABOUT_IT_URL);
-      expect(resp.text).toContain(saveAndContinueButtonText);
+      expect(resp.text).toContain(TRUST_INVOLVED_TITLE);
+      expect(resp.text).toContain(mockTrustData.trustName);
       expect(resp.text).not.toContain(PAGE_TITLE_ERROR);
     });
 
     test('when feature flag is off, 404 is returned', async () => {
       mockIsActiveFeature.mockReturnValue(false);
 
-      const resp = await request(app).get(UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL);
+      const resp = await request(app).get(pageUrl);
 
       expect(resp.status).toEqual(404);
       expect(resp.text).toContain(PAGE_NOT_FOUND_TEXT);
@@ -72,7 +81,7 @@ describe('Update - Trusts - Individuals or entities involved', () => {
     test('when feature flag is on, redirect to trusts associated with the entity page', async () => {
       mockIsActiveFeature.mockReturnValue(true);
 
-      const resp = await request(app).post(UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL);
+      const resp = await request(app).post(pageUrl).send({ noMoreToAdd: 'noMoreToAdd' });
 
       expect(resp.status).toEqual(302);
       expect(resp.header.location).toEqual(UPDATE_TRUSTS_ASSOCIATED_WITH_THE_OVERSEAS_ENTITY_URL);
@@ -81,7 +90,7 @@ describe('Update - Trusts - Individuals or entities involved', () => {
     test('when feature flag is off, 404 is returned', async () => {
       mockIsActiveFeature.mockReturnValue(false);
 
-      const resp = await request(app).post(UPDATE_TRUSTS_INDIVIDUALS_OR_ENTITIES_INVOLVED_URL);
+      const resp = await request(app).post(pageUrl);
 
       expect(resp.status).toEqual(404);
       expect(resp.text).toContain(PAGE_NOT_FOUND_TEXT);
