+++ views/includes/page-templates/trust-involved.html
@@ -1,0 +1,193 @@
+{% block content %}
+
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+
+      {% include "includes/list/errors.html" %}
+
+      <span class="govuk-caption-l">{{ pageData.trustData.trustName }}</span>
+      <h1 class="govuk-heading-xl">{{ pageParams.title }}</h1>
+
+      {% set boItems = [] %}
+      {% for bo in pageData.boInTrust %}
+        {% set type = pageData.beneficialOwnerTypeTitle[bo.type] %}
+        {% set boItems = (
+          boItems.push({
+            key: {
+              text: type
+            },
+            value: {
+              text: bo.name
+            }
+          }),
+          boItems
+        ) %}
+      {% endfor %}
+
+      {% set formerTrustees = [] %}
+      {% for trustee in pageData.formerTrusteeData %}
+        {% set type = pageData.trusteeTypeTitle[pageData.trusteeType.HISTORICAL] %}
+        {% if trustee.forename and trustee.surname%}
+          {% set name = trustee.forename + " " + trustee.surname %}
+          {% set formerTrusteeChangeName = OE_CONFIGS.INDIVIDUAL_NAME %}
+        {% else %}
+          {% set name = trustee.corporate_name %}
+          {% set formerTrusteeChangeName = OE_CONFIGS.CORPORATE_NAME %}
+        {% endif %}
+        {% set formerTrustees = (
+          formerTrustees.push({
+            key: {
+              text: type  
+            },
+            value: {
+              text: name
+            },
+            actions: {
+              items: [ CREATE_CHANGE_LINK(
+                OE_CONFIGS.TRUST_HISTORICAL_BENEFICIAL_OWNER_PAGE + "/" + trustee.id + formerTrusteeChangeName,
+                "Former Trustee" + trustee.name,
+                "change-former-trustee-button"
+             ) ]
+            }
+          }),
+          formerTrustees
+        ) %}
+      {% endfor %}
+
+      {% set individualTrustees = [] %}
+      {% for trustee in pageData.individualTrusteeData %}
+        {% set type = pageData.trusteeType.INDIVIDUAL | capitalize %}
+        {% set name = trustee.forename + " " + trustee.surname %}
+        {% set individualTrustees = (
+          individualTrustees.push({
+            key: {
+              text: type  
+            },
+            value: {
+              text: name
+            },
+            actions: {
+              items: [ CREATE_CHANGE_LINK(
+                OE_CONFIGS.TRUST_INDIVIDUAL_BENEFICIAL_OWNER_PAGE + "/" + trustee.id + OE_CONFIGS.FORENAME,
+                "Individual Trustee " + name,
+                "change-individual-trustee-button"
+             ) ]
+            }
+          }),
+          individualTrustees
+        ) %}
+      {% endfor %}
+
+      {% set legalEntityTrustees = [] %}
+      {% for trustee in pageData.trustees %}
+        {% set type = pageData.trusteeTypeTitle[trustee.trusteeItemType] %}
+        {% set legalEntityTrustees = (
+          legalEntityTrustees.push({
+            key: {
+              text: type
+            },
+            value: {
+              text: trustee.name
+            },
+            actions: {
+              items: [ CREATE_CHANGE_LINK(
+                OE_CONFIGS.TRUST_LEGAL_ENTITY_BENEFICIAL_OWNER_PAGE + "/" + trustee.id + OE_CONFIGS.LEGAL_ENTITY_NAME,
+                "Legal Entity Trustee " + trustee.name,
+                "change-legal-entity-trustee-button"
+             ) ]
+            }
+          }),
+          legalEntityTrustees
+        ) %}
+      {% endfor %}
+
+      {% if boItems | length or formerTrustees | length or individualTrustees | length or legalEntityTrustees | length %}
+        <h1 class="govuk-heading-m">
+          What <span class="govuk-visually-hidden">individuals or entities</span>
+          you have added so far for this trust
+        </h1>
+        {{ govukSummaryList({
+          rows: boItems
+        }) }}
+
+        {{ govukSummaryList({
+          rows: formerTrustees
+        }) }}
+
+        {{ govukSummaryList({
+          rows: individualTrustees
+        }) }}
+
+        {{ govukSummaryList({
+          rows: legalEntityTrustees
+        }) }}
+        
+      {% endif %}
+      
+
+      <form class="form" method="post">
+        {% set fieldParam = {
+          name: 'typeOfTrustee',
+          label: 'Which type of individual or entity do you want to add to this trust?',
+          error: errors.typeOfTrustee,
+          items: [
+            {
+              value: pageData.trusteeType.HISTORICAL,
+              text: "Former beneficial owner",
+              hint: {
+                text: "This is any individual or legal entity that was previously a registrable beneficial owner of the overseas entity, due to being a trustee of this trust."
+              },
+              checked: typeOfTrustee == pageData.trusteeType.HISTORICAL,
+              attributes: {
+                "data-event-id": "former-beneficial-owner-radio-option"
+              }
+            }, {
+              value: pageData.trusteeType.INDIVIDUAL,
+              text: "Individual",
+              hint: {
+                text: "This is an individual who is currently a beneficiary, settlor, grantor or interested person for this trust."
+              },
+               checked: typeOfTrustee == pageData.trusteeType.INDIVIDUAL,
+              attributes: {
+                "data-event-id": "individual-radio-option"
+              }
+            }, {
+              value: pageData.trusteeType.LEGAL_ENTITY,
+              text: "Legal entity",
+              hint: {
+                text: "This is a legal entity that is currently a beneficiary, settlor, grantor or interested person for this trust. You do not need to tell us about government or public authorities."
+              },
+              checked: typeOfTrustee == pageData.trusteeType.LEGAL_ENTITY,
+              attributes: {
+                "data-event-id": "legal-entity-radio-option"
+              }
+            }
+          ]
+        } %}
+        {% include "includes/inputs/radio-input.html" %}
+
+        <div class="govuk-button-group">
+          {{ govukButton({
+              text: "Add",
+            attributes: {
+              "id": "add",
+              "data-event-id": "trust-involved-page-add-button"
+            }
+            }) }}
+
+          {{ govukButton({
+              attributes: {
+                "id": "noMoreToAdd",
+                "name": "noMoreToAdd",
+                "value": "noMoreToAdd"
+              },
+              html: 'No more <span class="govuk-visually-hidden">individuals or entities</span> to add',
+              classes: "govuk-button--secondary"
+            }) }}
+        </div>
+
+      </form>
+    </div>
+  </div>
+
+{% endblock %}
